Separate tests by requirement and not by screen
